requirements.txt
arrow==0.14.5
binaryornot==0.4.4
biplist==1.0.3
boto3==1.9.213
botocore==1.12.213
briefcase==0.2.9
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
dmgbuild==1.3.2
docutils==0.15.2
ds-store==1.1.2
future==0.17.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
mac-alias==2.0.7
MarkupSafe==1.1.1
numpy==1.17.1
poyo==0.5.0
PyJWT==1.7.1
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rubicon-objc==0.3.1
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.6
toga==0.3.0.dev13
toga-cocoa==0.3.0.dev13
toga-core==0.3.0.dev13
travertino==0.1.2
twilio==6.29.3
urllib3==1.23
voc==0.1.6
whichcraft==0.6.0
